Удаляет обработчики события
